More secure and extensible
Now a Ticket contains the owner of the data, nodes from where we can fetch the data and the hash and format of the data.
Now the RawConversation and the RawMessage when are stored are signed.
Now the RawConversation and the RawMessage when is downloaded is verified with the owner public key.
Now we can add more providers to a Ticket.