Audio conversations.
There are really big performance problems.

So currently using the opus codec with default settings using 48000 Sample Rate and mono channel an opus packet has 960 samples.
So 50 Audio Frames will be sent per second.
An Audio Frame first needs to be stored in the iroh store, then a message with that ticket will be created,
then that message also needs to be stored and that message ticket will be sent using the protocol.
Then that message will be downloaded, parse the message if is a command, then if is a `/auto play {ticket}` will download the blob,
after that that blob will be sent to the decoder.

The problem that I see is that there are 3 downloads per Audio Frame, so 150 downloads per seccond.
But the performance problem looks like is on the CPU side, because CPU is used more the 150%

I know that this is a bad design for real time conversation, but I want everything to be archivabile.
And how the Audio Frames are stored, are a cryptografical proof by who sent them.