Implement TTL for tickets, messages and data will be deleted if their ticket exipres. TTL is aproximated to seconds.
TTL 0 means that a ticket don't expires.
Default TTL for a message is 0.
Default TTL for a audio stream is 10, so the audio data will have TTL 10 the stop message and the play message.

We don't have infinite memory, so the streams will not be stored for more then 10 seconds.

When a message is deleted the conversation tree is not updated, but because the conversation tree is not stored,
when relaunching the node the conversation tree will be build correctly.

But because the conversation tree is not updated we will used more and more memory, even when the messages and data is deleted.
Before 40MB per second, now 1MB per second, for 1 audio stream.